requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-ckeditor==5.7.1
django-geoposition==0.3.0
django-google-maps==0.11.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-pagination==1.0.7
django-phone-field==1.8.0
django-rest-auth==0.9.5
django-richtextfield==1.4.0
djangorestframework==3.10.3
gunicorn==20.0.0
heroku==0.1.4
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==1.5
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4

